requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
PyJWT==2.8.0
pytz==2024.1
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
twilio==9.0.0
typing_extensions==4.10.0
urllib3==2.2.1
yarl==1.9.4
